LPD-72213 Restructure folders
